Base Pothole Application
This is the base Pothole Application - collects accelerometer, gyroscope, and magnetometer data and prints them out in console. Code is commented and explained. Also, certain code blocks are commented out because that is based on the finalized algorithm - I just put in base numbers. Note, this version does not contain a current location file.